0073 prop_webtransport.rs をディレクトリモジュールに分割し PBT 間の重複テストを削除する
- prop_webtransport.rs (1802 行) を prop_webtransport/ ディレクトリモジュールに分割する
- capsule / connect / error / session / settings / stream の 6 サブモジュールに分類する
- prop_capsule.rs と重複する Capsule ラウンドトリップ 6 件を削除する
- prop_datagram.rs と重複する Datagram テスト 5 件を削除する
- .unwrap() を .expect("MESSAGE") に置き換え CLAUDE.md 規約に準拠する
